requirements.txt
discord.py==2.0.1
Pillow==9.3.0
praw==7.6.0
requests==2.28.1
opencv-python==4.6.0.66